solver works
* bound_type is extended to let it have summ
* ui connected
* generator is added
* solver has large pascal triangle to make sure free bombs calculated
* equations cache is added
* specter is used to calculate probablities one run
* total bombs count is used
* topology provided a way to iterate all cells
* tests are updaed

+++ src/engine/Equations.hpp
@@ -89,7 +89,7 @@ std::ostream& operator<<(std::ostream& o, const boost::numeric::ublas::vector<T>
  * x in [0, ..., bound]
  */
 struct Equations {
-    using bound_type = int8_t;
+    using bound_type = int64_t;
     using value_type = boost::rational<int64_t>;
     using matrix_type = boost::numeric::ublas::matrix<value_type>;
     using vector_type = boost::numeric::ublas::vector<value_type>;
